feat(profile): add non-interactive auth mode for coding agents (#157)
* feat(profile): add non-interactive auth mode for coding agents

* chore: update packages.lock.json after rebase on beta

* fix: resolve all new qodana findings from PR and beta rebase
CA1305, CA1031, InconsistentNaming, UnusedMember, MemberCanBeProtected,
UsingStatementResourceInitialization, S3358, MergeIntoPattern, CA1859,
RedundantUsingDirective, S5445, ReturnTypeCanBeNotNullable,
UnusedAutoPropertyAccessor, S1144, CoVariantArrayConversion,
AsyncMethodWithoutAwait, RedundantCast — all fixed in source and tests.
NonInteractive property is now used in PowerLogic.ExecuteAsync to propagate
DGTP_NON_INTERACTIVE for mid-command token refresh scenarios.
AuthCheckCommand ClassNeverInstantiated suppressed via ReSharper comment.

* docs: forbid manual edits to baseline.sarif.json in agent guidelines

+++ tests/dgt.power.push.tests/Logic/WebresourcesProcessorTests.cs
@@ -87,7 +87,7 @@ public class WebresourcesProcessorTests : PushTestsBase<PushCommand>
         var existingWebResource = new WebResource(existingWebResourceId)
         {
             Name = WebresourceLogicalName,
-            WebResourceType = new OptionSetValue((int)WebResource.Options.WebResourceType.ScriptJScript),
+            WebResourceType = new OptionSetValue(WebResource.Options.WebResourceType.ScriptJScript),
             Content = Convert.ToBase64String("old content"u8.ToArray())
         };
 
@@ -128,7 +128,7 @@ public class WebresourcesProcessorTests : PushTestsBase<PushCommand>
         var existingWebResource = new WebResource(existingWebResourceId)
         {
             Name = WebresourceLogicalName,
-            WebResourceType = new OptionSetValue((int)WebResource.Options.WebResourceType.ScriptJScript),
+            WebResourceType = new OptionSetValue(WebResource.Options.WebResourceType.ScriptJScript),
             Content = KnownJsBase64
         };
